Moderate settlement task scores
09.02.2023

Note: this commit primarily addresses the insufficient/inaccurate task
score for tending crop

- change: rework getTendingScore() in Crop to let the value of
          currentWorkRequired driving the settlement task score
- change: reduce the value of score in ToggleResourceProcessMeta

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/structure/building/function/farming/Farming.java
@@ -1216,8 +1216,8 @@ public class Farming extends Function {
 	 * 
 	 * @return
 	 */
-	public int getTendingScore() {
-		int score = 0;
+	public double getTendingScore() {
+		double score = 0;
 		for (Crop c : cropList) {
 			score += c.getTendingScore();
 		}
